Removed commands from index.js to external folder commands, swapi is called now within the whtb command and added embedded message
